Admin UI

+++ Backend/src/TaskManagement.Application/DTOs/WorkTask/UpdateWorkTaskDto.cs
@@ -7,7 +7,7 @@ namespace TaskManagement.Application.DTOs.WorkTask
         public string Title { get; set; } = string.Empty;
         public string? Description { get; set; }
         public int? Priority { get; set; }
-        public int? StoryPoints { get; set; }
+        public double? StoryPoints { get; set; }
         public Guid? AssignedUserId { get; set; }
         public DateTime? PlannedStartDate { get; set; }
         public DateTime? PlannedEndDate { get; set; }
